refactor: Add new repository package, for separation of responsibilities
