+++ app/src/main/java/br/com/jonatas/metronomeplus/data/repository/MeasureRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package br.com.jonatas.metronomeplus.data.provider
+package br.com.jonatas.metronomeplus.data.repository
 
 import br.com.jonatas.metronomeplus.domain.model.Beat
 import br.com.jonatas.metronomeplus.domain.model.BeatState
