+++ app/src/main/java/br/com/jonatas/metronomeplus/presenter/ui/viewmodel/MetronomeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import br.com.jonatas.metronomeplus.data.provider.AssetProvider
 import br.com.jonatas.metronomeplus.data.provider.AudioSettingsProvider
-import br.com.jonatas.metronomeplus.data.provider.MeasureRepository
+import br.com.jonatas.metronomeplus.data.repository.MeasureRepository
 import br.com.jonatas.metronomeplus.domain.engine.MetronomeEngine
 import br.com.jonatas.metronomeplus.domain.model.Measure
 import kotlinx.coroutines.CoroutineDispatcher
