+++ app/src/test/java/br/com/jonatas/metronomeplus/presenter/ui/viewmodel/MetronomeViewModelTest.kt
@@ -3,7 +3,7 @@ package br.com.jonatas.metronomeplus.presenter.ui.viewmodel
 import android.content.res.AssetManager
 import br.com.jonatas.metronomeplus.data.provider.AssetProvider
 import br.com.jonatas.metronomeplus.data.provider.AudioSettingsProvider
-import br.com.jonatas.metronomeplus.data.provider.MeasureRepository
+import br.com.jonatas.metronomeplus.data.repository.MeasureRepository
 import br.com.jonatas.metronomeplus.domain.engine.MetronomeEngine
 import br.com.jonatas.metronomeplus.domain.model.Beat
 import br.com.jonatas.metronomeplus.domain.model.BeatState
